flow optimizations
