+++ app/src/main/java/com/jys/catsapp/core/common/generic/usecase/BaseUseCase.kt
@@ -1,4 +1,4 @@
-package com.jys.catsapp.core.generic.usecase
+package com.jys.catsapp.core.common.generic.usecase
 
 interface BaseUseCase<In, Out>{
     suspend fun execute(input: In): Out
