+++ app/src/main/java/com/jys/catsapp/core/common/ui/VideoBackgroundScreen.kt
@@ -1,4 +1,4 @@
-package com.jys.catsapp.core.common
+package com.jys.catsapp.core.common.ui
 
 
 import android.net.Uri
