+++ app/src/main/java/com/jys/catsapp/core/common/utils/ConstantsAsset.kt
@@ -1,0 +1,5 @@
+package com.jys.catsapp.core.common.utils
+
+object ConstantsAsset {
+    const val VIDEO_SAMURAI_CAT = "asset:///samurai_cat.mp4"
+}
