+++ app/src/main/java/com/jys/catsapp/core/common/utils/DimensUtil.kt
@@ -1,4 +1,4 @@
-package com.jys.catsapp.core.common
+package com.jys.catsapp.core.common.utils
 
 import androidx.compose.ui.unit.dp
