+++ app/src/main/java/com/jys/catsapp/data/datasource/PexelPagingDataSource.kt
@@ -2,9 +2,9 @@ package com.jys.catsapp.data.datasource
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
-import com.jys.catsapp.core.generic.error.Result
-import com.jys.catsapp.core.generic.error.safeApiCall
-import com.jys.catsapp.core.utils.ConstantsUtil.PagingConstants.PAGE_SIZE
+import com.jys.catsapp.core.common.generic.error.Result
+import com.jys.catsapp.core.common.generic.error.safeApiCall
+import com.jys.catsapp.core.common.utils.ConstantsUtil.PagingConstants.PAGE_SIZE
 import com.jys.catsapp.data.network.PexelsApiService
 import com.jys.catsapp.data.network.model.Photo
