+++ app/src/main/java/com/jys/catsapp/di/NetworkModule.kt
@@ -1,8 +1,8 @@
 package com.jys.catsapp.di
 
 import com.jys.catsapp.BuildConfig
-import com.jys.catsapp.core.utils.ConstantsUtil.PagingConstants.PAGING_3_SERVICE
-import com.jys.catsapp.core.utils.ConstantsUtil.ServiceConstants.URL_PEXEL
+import com.jys.catsapp.core.common.utils.ConstantsUtil.PagingConstants.PAGING_3_SERVICE
+import com.jys.catsapp.core.common.utils.ConstantsUtil.ServiceConstants.URL_PEXEL
 import com.jys.catsapp.data.network.PexelsApiService
 import com.jys.catsapp.data.network.RetrofitObject
 import org.koin.core.qualifier.named
