+++ app/src/main/java/com/jys/catsapp/domain/usecase/GetCatUseCase.kt
@@ -1,8 +1,8 @@
 package com.jys.catsapp.domain.usecase
 
 import androidx.paging.PagingData
-import com.jys.catsapp.core.generic.usecase.BaseUseCase
-import com.jys.catsapp.core.utils.ConstantsUtil.PagingConstants.QUERY_CAT
+import com.jys.catsapp.core.common.generic.usecase.BaseUseCase
+import com.jys.catsapp.core.common.utils.ConstantsUtil.PagingConstants.QUERY_CAT
 import com.jys.catsapp.data.network.model.Photo
 import com.jys.catsapp.domain.repository.PexelPagingRepositoryInterface
 import kotlinx.coroutines.flow.Flow
