+++ app/src/main/java/com/jys/catsapp/domain/usecase/GetCatUseCaseWithRoom.kt
@@ -2,8 +2,8 @@ package com.jys.catsapp.domain.usecase
 
 import androidx.paging.PagingData
 import androidx.paging.map
-import com.jys.catsapp.core.generic.usecase.BaseUseCase
-import com.jys.catsapp.core.utils.ConstantsUtil.PagingConstants.QUERY_CAT
+import com.jys.catsapp.core.common.generic.usecase.BaseUseCase
+import com.jys.catsapp.core.common.utils.ConstantsUtil.PagingConstants.QUERY_CAT
 import com.jys.catsapp.domain.model.PhotoDomain
 import com.jys.catsapp.domain.model.toDomain
 import com.jys.catsapp.domain.repository.PexelPagingRepositoryInterface
