+++ app/src/main/java/com/jys/catsapp/navigation/Navigation.kt
@@ -1,20 +1,15 @@
 package com.jys.catsapp.navigation
 
-import androidx.compose.animation.EnterTransition
-import androidx.compose.animation.ExitTransition
-import androidx.compose.animation.core.tween
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.animation.slideInHorizontally
-import androidx.compose.animation.slideOutHorizontally
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.jys.catsapp.core.common.Constants.VIDEO_SAMURAI_CAT
-import com.jys.catsapp.core.common.Integers.Int1000
-import com.jys.catsapp.core.common.Integers.Int500
-import com.jys.catsapp.core.common.VideoBackgroundScreen
+import com.jys.catsapp.core.common.animation.defaultEnterTransition
+import com.jys.catsapp.core.common.animation.defaultExitTransition
+import com.jys.catsapp.core.common.animation.defaultPopEnterTransition
+import com.jys.catsapp.core.common.animation.defaultPopExitTransition
+import com.jys.catsapp.core.common.ui.VideoBackgroundScreen
+import com.jys.catsapp.core.common.utils.ConstantsAsset.VIDEO_SAMURAI_CAT
 import com.jys.catsapp.navigation.NavigationRoute.DASHBOARD
 import com.jys.catsapp.navigation.NavigationRoute.HOME_WITHOUT_ROOM
 import com.jys.catsapp.navigation.NavigationRoute.HOME_WITH_ROOM
@@ -50,18 +45,3 @@ fun Navigation() {
 
 
 
-fun defaultEnterTransition(): EnterTransition {
-    return slideInHorizontally(initialOffsetX = { Int1000 }) + fadeIn(animationSpec = tween(Int500))
-}
-
-fun defaultExitTransition(): ExitTransition {
-    return slideOutHorizontally(targetOffsetX = { -Int1000 }) + fadeOut(animationSpec = tween(Int500))
-}
-
-fun defaultPopEnterTransition(): EnterTransition {
-    return slideInHorizontally(initialOffsetX = { -Int1000 }) + fadeIn(animationSpec = tween(Int500))
-}
-
-fun defaultPopExitTransition(): ExitTransition {
-    return slideOutHorizontally(targetOffsetX = { Int1000 }) + fadeOut(animationSpec = tween(Int500))
-}
