+++ app/src/main/java/com/jys/catsapp/presentation/HomeScreen.kt
@@ -1,15 +1,18 @@
 package com.jys.catsapp.presentation
 
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
 import androidx.paging.PagingData
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
-import com.jys.catsapp.core.mock.mockValues
+import com.jys.catsapp.core.common.mock.mockValues
 import com.jys.catsapp.data.network.model.Photo
 import com.jys.catsapp.presentation.components.CatListComponent
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -30,7 +33,7 @@ private fun HomeScreenContent(
     stateUI: LazyPagingItems<Photo>
 ) {
     Scaffold(
-        modifier = Modifier.fillMaxSize(),) { paddingValues ->
+        modifier = Modifier.fillMaxSize()) { paddingValues ->
         CatListComponent(
             modifier = Modifier,
             paddingValues = paddingValues,
